Remove warnings on MacOS in Clang

+++ src/properties/scalar/ulongproperty.cpp
@@ -71,14 +71,12 @@ namespace properties {
         return true;                                                                     \
     }
 
-REGISTER_NUMERICALPROPERTY_SOURCE(ULongProperty, unsigned long, (unsigned long)0,
+REGISTER_NUMERICALPROPERTY_SOURCE(ULongProperty, unsigned long, 0ul,
                                   numeric_limits<unsigned long>::lowest(),
-                                  numeric_limits<unsigned long>::max(), (unsigned long)1,
-                                  DEFAULT_FROM_LUA_LAMBDA(unsigned long,
-                                                          (unsigned long)(0)),
+                                  numeric_limits<unsigned long>::max(), 1ul,
+                                  DEFAULT_FROM_LUA_LAMBDA(unsigned long, 0ul),
                                   DEFAULT_TO_LUA_LAMBDA(unsigned long),
-                                  DEFAULT_FROM_STRING_LAMBDA(unsigned long,
-                                                          (unsigned long)(0)),
+                                  DEFAULT_FROM_STRING_LAMBDA(unsigned long, 0ul),
                                   DEFAULT_TO_STRING_LAMBDA(unsigned long),
                                   LUA_TNUMBER);
